bot/requirements.txt
discord-py
python-dotenv
matplotlib
requests
aiohttp
asyncio
supabase